seperated the doctor and patient routes to seperate files.
