+++ controllers/profiles/admin.js
@@ -1,0 +1,5 @@
+// IMPORTED MODULES
+const express = require("express");
+const router = express.Router();
+
+module.exports = router;
