+++ helper/createToken.js
@@ -4,7 +4,7 @@ const jwt = require("jsonwebtoken");
 const createToken = (user) => {
   const token = user._id
     ? jwt.sign({ id: user._id }, process.env.JWT_SECRET)
-    : jwt.sign({ user }, process.env.JWT_SECRET);
+    : jwt.sign({ type: user }, process.env.JWT_SECRET);
   return token;
 };
